Fixed All warnings. Created Debug Class with Static Log(std::string) and LogError(std::string) to make debug logging easier. (And to have errors light up red more consistently in map editor) Player Now Dies when colliding with another object while struggling Map Editor Console Seems to recieve all the logs in one bunch instead of line by line. Turkey Now Walks ON the street Shadow is now softer. Turkey Cage now gives 100 Coins

+++ Game/_vs2015/ScoreBoard.cpp
@@ -35,7 +35,7 @@ void ScoreBoard::UpdateScores() {
 }
 
 void ScoreBoard::OrganizeScreen() {
-	_background->sprite.setPosition(_background->sprite.getTexture()->getSize().x / 2, _background->sprite.getTexture()->getSize().y / 2);
+	_background->sprite.setPosition((float)_background->sprite.getTexture()->getSize().x / 2.0f, _background->sprite.getTexture()->getSize().y / 2.0f);
 	
 	float wholeBlockOffset = 380;
 	float lineOffset = 47.5f;
